feat: Add comprehensive Swagger documentation and fix User address foreign keys
BREAKING CHANGE: User address identifier fields changed from Integer to Long

Changes:
- Update User entity: address identifiers Integer -> Long (BIGINT)
- Add comprehensive @Schema annotations to all User DTOs for API documentation
- Update UserCreateDTO, UserUpdateDTO, UserDTO with complete Swagger specs
- Create V5.0 migration: ALTER columns to BIGINT and ADD FK constraints
- Update Postman collection with address identifier fields
- Fix SecurityConfig to allow Swagger UI access
- Add comprehensive API testing documentation

Database Migration V5.0:
- Change Residential_Address_Identifier and Postal_Address_Identifier to BIGINT
- Add foreign key constraints (fk_add1_id, fk_add2_id) with ON DELETE SET NULL
- Idempotent migration with conditional constraint creation

API Documentation:
- All User DTO fields now have descriptions, examples, and type specifications
- Address identifiers properly documented as int64 (Long) with FK references
- Complete Swagger UI support for UI development

Testing:
- Updated Postman collection with address identifier examples
- Added automated test script (test-swagger-ui.ps1)
- Created comprehensive API testing guide

Documentation:
- API_TESTING_GUIDE.md - Complete Swagger UI and Postman usage guide
- DATA_MODEL_ANALYSIS.md - Entity schema comparison
- SCHEMA_ALIGNMENT_REPORT.md - Schema alignment verification
- UPDATE_SUMMARY.md - Comprehensive change summary

Benefits:
- Type safety: Long matches Address entity BIGINT primary key
- Data integrity: FK constraints enforce referential integrity
- Self-documenting API: Complete Swagger annotations
- Better DX: Clear examples and descriptions for all fields

+++ src/main/java/com/fincore/usermgmt/entity/User.java
@@ -43,10 +43,10 @@ public class User {
     private LocalDate dateOfBirth;
 
     @Column(name = "Residential_Address_Identifier")
-    private Integer residentialAddressIdentifier;
+    private Long residentialAddressIdentifier;
 
     @Column(name = "Postal_Address_Identifier")
-    private Integer postalAddressIdentifier;
+    private Long postalAddressIdentifier;
 
     @Column(name = "Status_Description", length = 20)
     private String statusDescription;
